Added checking email input to resolve issue #612 (#663)
Closes issue #612 @the-dagger. Please do check it out and merge.

I added code to check for the following:
     No spacial characters before and after @
     (-) and (.) should not be together after @
     2 characters must before @ 
     Email length should be less 128 characters
{Follows regex validation technique}

How it works,
On press of the generate button the code checks if the email is a regular email address or not with the help of the checkemail() function. If not reload the page as in case if there was nothing was there as input else proceed and generate the apk file!